Rename modules to align terminology (validation instead of validate/validators)
